Adding LL for euphonic.
missing: download of the data, proper testing,
detached app.
